+++ src/Transcoder.php
@@ -20,6 +20,7 @@ use craft\events\ModelEvent;
 use craft\events\PluginEvent;
 use craft\events\RegisterCacheOptionsEvent;
 use craft\events\RegisterUrlRulesEvent;
+use craft\events\TemplateEvent;
 use craft\helpers\Assets as AssetsHelper;
 use craft\helpers\FileHelper;
 use craft\helpers\UrlHelper;
@@ -28,6 +29,7 @@ use craft\services\Plugins;
 use craft\utilities\ClearCaches;
 use craft\web\twig\variables\CraftVariable;
 use craft\web\UrlManager;
+use craft\web\View;
 use nystudio107\transcoder\models\Settings;
 use nystudio107\transcoder\jobs\EncodeVideo;
 use nystudio107\transcoder\services\ServicesTrait;
@@ -101,6 +103,8 @@ class Transcoder extends Plugin
         $this->addComponents();
         // Install our global event handlers
         $this->installEventHandlers();
+        // Register settings page tabs
+        $this->registerSettingsTabs();
         // We've loaded!
         Craft::info(
             Craft::t(
@@ -157,6 +161,38 @@ class Transcoder extends Plugin
         return Craft::$app->getView()->renderTemplate('transcoder/settings', [
             'settings' => $this->getSettings(),
         ]);
+    }
+
+    /**
+     * Register Craft CP tabs for the plugin settings page.
+     */
+    protected function registerSettingsTabs(): void
+    {
+        Event::on(
+            View::class,
+            View::EVENT_BEFORE_RENDER_TEMPLATE,
+            function(TemplateEvent $event) {
+                if (
+                    $event->template === 'settings/plugins/_settings.twig'
+                    && ($event->variables['plugin']->handle ?? null) === $this->handle
+                ) {
+                    $event->variables['tabs'] = [
+                        [
+                            'label' => Craft::t('transcoder', 'Video queue'),
+                            'url' => '#settings-tab-video-queue',
+                        ],
+                        [
+                            'label' => Craft::t('transcoder', 'Video posters'),
+                            'url' => '#settings-tab-video-posters',
+                        ],
+                        [
+                            'label' => Craft::t('transcoder', 'Video watermark'),
+                            'url' => '#settings-tab-video-watermark',
+                        ],
+                    ];
+                }
+            }
+        );
     }
 
     /**